+++ src/pydvl/utils/parallel/backend.py
@@ -1,21 +1,10 @@
 from __future__ import annotations
 
+import logging
 import os
-from abc import ABCMeta, abstractmethod
+from abc import abstractmethod
 from concurrent.futures import Executor
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    Iterable,
-    List,
-    Optional,
-    Tuple,
-    Type,
-    TypeVar,
-    Union,
-    cast,
-)
+from typing import Any, Callable, Iterable, Type, TypeVar, cast
 
 import joblib
 import ray
@@ -25,42 +14,22 @@ from ray import ObjectRef
 from ray.util.joblib import register_ray
 
 from ..config import ParallelConfig
+from ..types import NoPublicConstructor
+from .futures.ray import CancellationPolicy
 
 __all__ = ["init_parallel_backend", "effective_n_jobs", "available_cpus"]
 
+
 T = TypeVar("T")
 
-
-class NoPublicConstructor(ABCMeta):
-    """Metaclass that ensures a private constructor
-
-    If a class uses this metaclass like this:
-
-        class SomeClass(metaclass=NoPublicConstructor):
-            pass
-
-    If you try to instantiate your class (`SomeClass()`),
-    a `TypeError` will be thrown.
-
-    Taken almost verbatim from:
-    https://stackoverflow.com/a/64682734
-    """
-
-    def __call__(cls, *args, **kwargs):
-        raise TypeError(
-            f"{cls.__module__}.{cls.__qualname__} cannot be initialized directly. "
-            "Use the proper factory instead."
-        )
-
-    def create(cls, *args: Any, **kwargs: Any):
-        return super().__call__(*args, **kwargs)
+log = logging.getLogger(__name__)
 
 
 class BaseParallelBackend(metaclass=NoPublicConstructor):
-    """Abstract base class for all parallel backends"""
+    """Abstract base class for all parallel backends."""
 
-    config: Dict[str, Any] = {}
-    BACKENDS: Dict[str, "Type[BaseParallelBackend]"] = {}
+    config: dict[str, Any] = {}
+    BACKENDS: dict[str, "Type[BaseParallelBackend]"] = {}
 
     def __init_subclass__(cls, *, backend_name: str, **kwargs):
         super().__init_subclass__(**kwargs)
@@ -72,8 +41,9 @@ class BaseParallelBackend(metaclass=NoPublicConstructor):
         cls,
         max_workers: int | None = None,
         config: ParallelConfig = ParallelConfig(),
-        **kwargs,
+        cancel_futures: CancellationPolicy = CancellationPolicy.PENDING,
     ) -> Executor:
+        """Returns an executor for the parallel backend."""
         ...
 
     @abstractmethod
@@ -127,9 +97,11 @@ class JoblibParallelBackend(BaseParallelBackend, backend_name="joblib"):
         cls,
         max_workers: int | None = None,
         config: ParallelConfig = ParallelConfig(),
-        **kwargs,
+        cancel_futures: CancellationPolicy = CancellationPolicy.NONE,
     ) -> Executor:
-        return cast(Executor, get_reusable_executor(max_workers=max_workers, **kwargs))
+        if cancel_futures is not CancellationPolicy.NONE:
+            log.warning("CancellationPolicy is not supported by the joblib backend")
+        return cast(Executor, get_reusable_executor(max_workers=max_workers))
 
     def get(self, v: T, *args, **kwargs) -> T:
         return v
@@ -146,7 +118,7 @@ class JoblibParallelBackend(BaseParallelBackend, backend_name="joblib"):
         """
         return delayed(fun)  # type: ignore
 
-    def wait(self, v: List[T], *args, **kwargs) -> Tuple[List[T], List[T]]:
+    def wait(self, v: list[T], *args, **kwargs) -> tuple[list[T], list[T]]:
         return v, []
 
     def _effective_n_jobs(self, n_jobs: int) -> int:
@@ -182,16 +154,14 @@ class RayParallelBackend(BaseParallelBackend, backend_name="ray"):
         cls,
         max_workers: int | None = None,
         config: ParallelConfig = ParallelConfig(),
-        **kwargs,
+        cancel_futures: CancellationPolicy = CancellationPolicy.PENDING,
     ) -> Executor:
-        from pydvl.utils.parallel.futures import RayExecutor
+        from pydvl.utils.parallel.futures.ray import RayExecutor
 
-        return RayExecutor(max_workers, config=config, **kwargs)  # type: ignore
+        return RayExecutor(max_workers, config=config, cancel_futures=cancel_futures)  # type: ignore
 
-    def get(
-        self, v: Union[ObjectRef, Iterable[ObjectRef], T], *args, **kwargs
-    ) -> Union[T, Any]:
-        timeout: Optional[float] = kwargs.get("timeout", None)
+    def get(self, v: ObjectRef | Iterable[ObjectRef] | T, *args, **kwargs) -> T | Any:
+        timeout: float | None = kwargs.get("timeout", None)
         if isinstance(v, ObjectRef):
             return ray.get(v, timeout=timeout)
         elif isinstance(v, Iterable):
@@ -199,7 +169,7 @@ class RayParallelBackend(BaseParallelBackend, backend_name="ray"):
         else:
             return v
 
-    def put(self, v: T, *args, **kwargs) -> Union["ObjectRef[T]", T]:
+    def put(self, v: T, *args, **kwargs) -> "ObjectRef[T]" | T:
         try:
             return ray.put(v, **kwargs)  # type: ignore
         except TypeError:
@@ -218,10 +188,10 @@ class RayParallelBackend(BaseParallelBackend, backend_name="ray"):
         return ray.remote(fun).remote  # type: ignore
 
     def wait(
-        self, v: List["ObjectRef"], *args, **kwargs
-    ) -> Tuple[List[ObjectRef], List[ObjectRef]]:
+        self, v: list[ObjectRef], *args, **kwargs
+    ) -> tuple[list[ObjectRef], list[ObjectRef]]:
         num_returns: int = kwargs.get("num_returns", 1)
-        timeout: Optional[float] = kwargs.get("timeout", None)
+        timeout: float | None = kwargs.get("timeout", None)
         return ray.wait(  # type: ignore
             v,
             num_returns=num_returns,
@@ -237,9 +207,7 @@ class RayParallelBackend(BaseParallelBackend, backend_name="ray"):
         return eff_n_jobs
 
 
-def init_parallel_backend(
-    config: ParallelConfig,
-) -> BaseParallelBackend:
+def init_parallel_backend(config: ParallelConfig) -> BaseParallelBackend:
     """Initializes the parallel backend and returns an instance of it.
 
     :param config: instance of :class:`~pydvl.utils.config.ParallelConfig`
